week04-homework/pyproject.toml
[project]
name = "week04-homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"python-dotenv",
	"langchain",
	"langchain-openai",
	"langchain-community",
	# add your dependencies here and run `uv sync` to install them
	"langchain-core>=1.1.0",
	"langgraph>=0.2.0",
	"fastapi>=0.100.0",
	"uvicorn>=0.23.0",
	"watchdog>=3.0.0",
	"pytest>=7.4.0",
	"httpx>=0.24.0",
]
